[TASK] Implement HTTP request to retrieve instance meta data
Meta data of an EC2 instance can be retrieved by accessing an
AWS-internal URI (169.254.169.254). The instance type (for example
"t2.large") is one of these meta data.

This change implements the HTTP request to retrieve the instance
type from the meta data and display the result in the System
Information Toolbar.